+++ configserver/src/main/java/com/yahoo/vespa/config/server/host/HostValidator.java
@@ -1,5 +1,5 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.config.server;
+package com.yahoo.vespa.config.server.host;
 
 import java.util.Collection;
