+++ configserver/src/main/resources/configserver-app/services.xml
@@ -19,7 +19,7 @@
     <component id="com.yahoo.vespa.config.server.InjectedGlobalComponentRegistry" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.tenant.Tenants" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.application.PermanentApplicationPackage" bundle="configserver" />
-    <component id="com.yahoo.vespa.config.server.HostRegistries" bundle="configserver" />
+    <component id="com.yahoo.vespa.config.server.host.HostRegistries" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.ApplicationRepository" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.version.VersionState" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.zookeeper.ConfigCurator" bundle="configserver" />
