+++ configserver/src/test/java/com/yahoo/vespa/config/server/rpc/TestWithRpc.java
@@ -8,13 +8,12 @@ import com.yahoo.jrt.Spec;
 import com.yahoo.jrt.Supervisor;
 import com.yahoo.jrt.Transport;
 import com.yahoo.vespa.config.GenerationCounter;
-import com.yahoo.vespa.config.server.HostRegistries;
+import com.yahoo.vespa.config.server.host.HostRegistries;
 import com.yahoo.vespa.config.server.MemoryGenerationCounter;
 import com.yahoo.vespa.config.server.PortRangeAllocator;
 import com.yahoo.vespa.config.server.SuperModelController;
 import com.yahoo.vespa.config.server.TestConfigDefinitionRepo;
 import com.yahoo.vespa.config.server.monitoring.Metrics;
-import com.yahoo.vespa.config.server.rpc.RpcServer;
 import com.yahoo.vespa.config.server.tenant.MockTenantProvider;
 import org.junit.After;
 import org.junit.Before;
